Update.
2003-04-22  Ulrich Drepper  <drepper@redhat.com>

	* nscd/nscd-client.h: Add declaration for __nscd_open_socket.
	* nscd/nscd_gethst_r.c (__nscd_open_socket): Renamed from
	open_socket.  Not static anymore.
	(nscd_gethst_r): Use __nscd_open_socket.
	* nscd/nscd_getgr_r.c (open_socket): Removed.
	(nscd_getgr_r): Use __nscd_open_socket.
	* nscd/nscd_getpw_r.c (open_socket): Removed.
	(nscd_getpw_r): Use __nscd_open_socket.

	* nscd/nscd.c (main): Change type of fdn to long int and use strtol.
	* nscd/connections.c (handle_request): Add cast to avoid warning.

+++ nscd/connections.c
@@ -296,7 +296,7 @@ cannot handle old request version %d; current version is %d"),
 	  /* No, sent the prepared record.  */
 	  if (TEMP_FAILURE_RETRY (write (fd, db->disabled_iov->iov_base,
 					 db->disabled_iov->iov_len))
-	      != db->disabled_iov->iov_len
+	      != (ssize_t) db->disabled_iov->iov_len
 	      && __builtin_expect (debug_level, 0) > 0)
 	    {
 	      /* We have problems sending the result.  */
